refactor: extract format_event to avoid circular import
